requirements.txt
faster-whisper>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0

